Use SurfaceRegistry instead of AppRegistry
Summary: Right now we render a surface by calling `AppRegistry.runApplication`, but the way we do this relies on the batched bridge. For Venice (bridgeless RN) I created a new module for registering a surface called SurfaceRegistry, which uses a global variable instead of registering itself as a callable module on the bridge. If that global variable exists, we can use that to start the surface instead of calling AppRegistry.

Reviewed By: sahrens

Differential Revision: D15502241

fbshipit-source-id: 0b8d4677f1df41f46d84444567a30e40e21fed3d